refactored stats
