+++ src/stats/output_formats/build-time-format.json
@@ -1,7 +1,5 @@
 {
   "diff": {
-    "cpu_gpu": {}
-  },
-  "best_performer": "",
-  "worst_performer": ""
+    "gpu_pipe": {}
+  }
 }
